+++ SystemUtilities.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.3.32519.111
+# Visual Studio Version 18
+VisualStudioVersion = 18.0.11018.127 d18.0
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Environment", "src\Environment\Environment.csproj", "{B9A3EBB2-A213-4E09-9A53-966D4441D0D9}"
 EndProject
@@ -18,6 +18,8 @@ EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common", "src\Common\Common.csproj", "{E5935615-15A5-46BC-AC10-F93B6C8E5923}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Common.Tests", "test\Common.Tests\Common.Tests.csproj", "{24C5ABDA-35C4-42E6-9528-3AFB613BD7F2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -53,6 +55,10 @@ Global
 		{E5935615-15A5-46BC-AC10-F93B6C8E5923}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E5935615-15A5-46BC-AC10-F93B6C8E5923}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E5935615-15A5-46BC-AC10-F93B6C8E5923}.Release|Any CPU.Build.0 = Release|Any CPU
+		{24C5ABDA-35C4-42E6-9528-3AFB613BD7F2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{24C5ABDA-35C4-42E6-9528-3AFB613BD7F2}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{24C5ABDA-35C4-42E6-9528-3AFB613BD7F2}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{24C5ABDA-35C4-42E6-9528-3AFB613BD7F2}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -61,6 +67,7 @@ Global
 		{6353FCF3-AC19-4BC1-A579-996034B3159E} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{32DB8B9A-7C0E-4109-B96F-6D77A04CEACC} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 		{8B47D7F6-BAF5-436E-994E-04C8581DC8F9} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{24C5ABDA-35C4-42E6-9528-3AFB613BD7F2} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {6D5F83DF-0D61-4D02-8686-CF7B16311BDB}
