+++ test/Environment.Tests/MockEnvironmentTestBase.cs
@@ -1,7 +1,5 @@
 ﻿using DevOptimal.SystemUtilities.Environment.Abstractions;
-using Microsoft.QualityTools.Testing.Fakes;
 using System;
-using System.Diagnostics.Fakes;
 
 namespace DevOptimal.SystemUtilities.Environment.Tests
 {
@@ -16,16 +14,6 @@ namespace DevOptimal.SystemUtilities.Environment.Tests
         public void TestInitialize()
         {
             environment = new MockEnvironment();
-        }
-
-        protected static IDisposable CreateShimsContext()
-        {
-            var context = ShimsContext.Create();
-
-            ShimProcess.AllInstances.IdGet = p => System.Environment.ProcessId + 1;
-            ShimProcess.AllInstances.StartTimeGet = p => DateTime.Now;
-
-            return context;
         }
     }
 }
