+++ test/FileSystem.Tests/StateManagement/FileSystemSnapshotterTests.cs
@@ -331,128 +331,6 @@ namespace DevOptimal.SystemUtilities.FileSystem.Tests.StateManagement
             });
         }
 
-        [TestMethod]
-        [TestCategory("OmitFromCI")] // Fakes require Visual Studio Enterprise, but agent machines only have Community installed.
-        public void RestoresAbandonedDirectorySnapshots()
-        {
-            using var restoreSnapshotter = CreateSnapshotter();
-            var directoryPath = @"C:\foo";
-
-            /*
-             * First, we will test restoring a directory that didn't exist when the snapshot was taken, but has since been created
-             */
-            // Delete the directory, if it exists
-            if (fileSystem.DirectoryExists(directoryPath))
-            {
-                fileSystem.DeleteDirectory(directoryPath, recursive: true);
-            }
-
-            // Simulate taking a snapshot of the directory from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotDirectory(directoryPath);
-            }
-
-            // Create the directory
-            fileSystem.CreateDirectory(directoryPath);
-
-            // Restore the snapshot
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the directory has been deleted
-            Assert.IsFalse(fileSystem.DirectoryExists(directoryPath));
-
-            /*
-             * Next, we will test restoring a directory that did exist when the snapshot was taken, but has since been deleted
-             */
-            // Create the directory
-            fileSystem.CreateDirectory(directoryPath);
-
-            // Simulate taking a snapshot of the directory from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotDirectory(directoryPath);
-            }
-
-            // Delete the directory
-            fileSystem.DeleteDirectory(directoryPath, recursive: true);
-
-            // Restore the snapshot
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the directory has been created
-            Assert.IsTrue(fileSystem.DirectoryExists(directoryPath));
-        }
-
-        [TestMethod]
-        [TestCategory("OmitFromCI")] // Fakes require Visual Studio Enterprise, but agent machines only have Community installed.
-        public void RestoresAbandonedFileSnapshots()
-        {
-            using var restoreSnapshotter = CreateSnapshotter();
-            var filePath = @"C:\foo\bar.txt";
-            var expectedFileBytes = Encoding.UTF8.GetBytes("Hello, world!");
-
-            /*
-             * First, we will test restoring a file that didn't exist when the snapshot was taken, but has since been created
-             */
-            // Delete the file, if it exists
-            if (fileSystem.FileExists(filePath))
-            {
-                fileSystem.DeleteFile(filePath);
-            }
-
-            // Simulate taking a snapshot of the file from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotFile(filePath);
-            }
-
-            // Create the file
-            using (var stream = fileSystem.OpenFile(filePath, FileMode.CreateNew, FileAccess.Write, FileShare.None))
-            {
-                stream.Write(expectedFileBytes);
-            }
-
-            // Restore the snapshot
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the file has been deleted
-            Assert.IsFalse(fileSystem.FileExists(filePath));
-
-            /*
-             * Next, we will test restoring a file that did exist when the snapshot was taken, but has since been deleted
-             */
-            // Create the file
-            using (var stream = fileSystem.OpenFile(filePath, FileMode.CreateNew, FileAccess.Write, FileShare.None))
-            {
-                stream.Write(expectedFileBytes);
-            }
-
-            // Simulate taking a snapshot of the file from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotFile(filePath);
-            }
-
-            // Delete the file
-            fileSystem.DeleteFile(filePath);
-
-            // Restore the snapshot
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the file has been created
-            var actualFileBytes = new byte[expectedFileBytes.Length];
-            using (var stream = fileSystem.OpenFile(filePath, FileMode.Open, FileAccess.Read, FileShare.None))
-            {
-                stream.ReadExactly(actualFileBytes, 0, expectedFileBytes.Length);
-            }
-            CollectionAssert.AreEqual(expectedFileBytes, actualFileBytes);
-        }
-
         #endregion
 
         private FileSystemSnapshotter CreateSnapshotter()
