+++ test/Registry.Tests/StateManagement/RegistrySnapshotterTests.cs
@@ -457,209 +457,6 @@ namespace DevOptimal.SystemUtilities.Registry.Tests.StateManagement
             });
         }
 
-        [TestMethod]
-        [TestCategory("OmitFromCI")] // Fakes require Visual Studio Enterprise, but agent machines only have Community installed.
-        public void RestoresAbandonedRegistryKeySnapshots()
-        {
-            using var restoreSnapshotter = CreateSnapshotter();
-            var registryHive = RegistryHive.LocalMachine;
-            var registryView = RegistryView.Default;
-            var registrySubKey = @"SOFTWARE\Microsoft\StrongName\Verification";
-
-            /*
-             * First, we will test restoring a registry key that didn't exist when the snapshot was taken, but has since been created
-             */
-            // Delete the registry key, if it exists
-            if (registry.RegistryKeyExists(registryHive, registryView, registrySubKey))
-            {
-                registry.DeleteRegistryKey(registryHive, registryView, registrySubKey, recursive: true);
-            }
-
-            // Simulate taking a snapshot of the registry key from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotRegistryKey(registryHive, registryView, registrySubKey);
-            }
-
-            // Create the registry key
-            registry.CreateRegistryKey(registryHive, registryView, registrySubKey);
-
-            // Restore the snapshot
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the registry key has been deleted
-            Assert.IsFalse(registry.RegistryKeyExists(registryHive, registryView, registrySubKey));
-
-            /*
-             * Next, we will test restoring a registry key that did exist when the snapshot was taken, but has since been deleted
-             */
-            // Create the registry key
-            registry.CreateRegistryKey(registryHive, registryView, registrySubKey);
-
-            // Simulate taking a snapshot of the registry key from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotRegistryKey(registryHive, registryView, registrySubKey);
-            }
-
-            // Delete the registry key
-            registry.DeleteRegistryKey(registryHive, registryView, registrySubKey, recursive: true);
-
-            // Restore the snapshot
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the registry key has been created
-            Assert.IsTrue(registry.RegistryKeyExists(registryHive, registryView, registrySubKey));
-        }
-
-        [TestMethod]
-        [TestCategory("OmitFromCI")] // Fakes require Visual Studio Enterprise, but agent machines only have Community installed.
-        public void RestoresAbandonedRegistryValueSnapshots()
-        {
-            using var restoreSnapshotter = CreateSnapshotter();
-            // Create registry key
-            var registryHive = RegistryHive.LocalMachine;
-            var registryView = RegistryView.Default;
-            var registrySubKey = @"SOFTWARE\Microsoft\StrongName\Verification";
-            registry.CreateRegistryKey(registryHive, registryView, registrySubKey);
-
-            // One registry key value for each supported type
-            var stringRegistryValueName = RegistryValueKind.String.ToString();
-            var stringRegistryValueExpectedValue = "foo";
-            var stringRegistryValueExpectedKind = RegistryValueKind.String;
-            var expandStringRegistryValueName = RegistryValueKind.ExpandString.ToString();
-            var expandStringRegistryValueExpectedValue = "bar";
-            var expandStringRegistryValueExpectedKind = RegistryValueKind.ExpandString;
-            var binaryRegistryValueName = RegistryValueKind.Binary.ToString();
-            var binaryRegistryValueExpectedValue = Encoding.UTF8.GetBytes("Hello, world!");
-            var binaryRegistryValueExpectedKind = RegistryValueKind.Binary;
-            var dwordRegistryValueName = RegistryValueKind.DWord.ToString();
-            var dwordRegistryValueExpectedValue = (int)3;
-            var dwordRegistryValueExpectedKind = RegistryValueKind.DWord;
-            var qwordRegistryValueName = RegistryValueKind.QWord.ToString();
-            var qwordRegistryValueExpectedValue = (long)3;
-            var qwordRegistryValueExpectedKind = RegistryValueKind.QWord;
-            var multiStringRegistryValueName = RegistryValueKind.MultiString.ToString();
-            var multiStringRegistryValueExpectedValue = new[] { "hello", "world" };
-            var multiStringRegistryValueExpectedKind = RegistryValueKind.MultiString;
-
-            /*
-             * First, we will test restoring values that didn't exist when the snapshot was taken, but have since been created
-             */
-            // Delete the registry values, if they exist
-            if (registry.RegistryValueExists(registryHive, registryView, registrySubKey, stringRegistryValueName))
-            {
-                registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, stringRegistryValueName);
-            }
-            if (registry.RegistryValueExists(registryHive, registryView, registrySubKey, expandStringRegistryValueName))
-            {
-                registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, expandStringRegistryValueName);
-            }
-            if (registry.RegistryValueExists(registryHive, registryView, registrySubKey, binaryRegistryValueName))
-            {
-                registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, binaryRegistryValueName);
-            }
-            if (registry.RegistryValueExists(registryHive, registryView, registrySubKey, dwordRegistryValueName))
-            {
-                registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, dwordRegistryValueName);
-            }
-            if (registry.RegistryValueExists(registryHive, registryView, registrySubKey, qwordRegistryValueName))
-            {
-                registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, qwordRegistryValueName);
-            }
-            if (registry.RegistryValueExists(registryHive, registryView, registrySubKey, multiStringRegistryValueName))
-            {
-                registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, multiStringRegistryValueName);
-            }
-
-            // Simulate taking snapshots of the registry values from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, stringRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, expandStringRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, binaryRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, dwordRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, qwordRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, multiStringRegistryValueName);
-            }
-
-            // Create the registry values
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, stringRegistryValueName, stringRegistryValueExpectedValue, stringRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, expandStringRegistryValueName, expandStringRegistryValueExpectedValue, expandStringRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, binaryRegistryValueName, binaryRegistryValueExpectedValue, binaryRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, dwordRegistryValueName, dwordRegistryValueExpectedValue, dwordRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, qwordRegistryValueName, qwordRegistryValueExpectedValue, qwordRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, multiStringRegistryValueName, multiStringRegistryValueExpectedValue, multiStringRegistryValueExpectedKind);
-
-            // Restore the snapshots
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the registry values have been deleted
-            Assert.IsFalse(registry.RegistryValueExists(registryHive, registryView, registrySubKey, stringRegistryValueName));
-            Assert.IsFalse(registry.RegistryValueExists(registryHive, registryView, registrySubKey, expandStringRegistryValueName));
-            Assert.IsFalse(registry.RegistryValueExists(registryHive, registryView, registrySubKey, binaryRegistryValueName));
-            Assert.IsFalse(registry.RegistryValueExists(registryHive, registryView, registrySubKey, dwordRegistryValueName));
-            Assert.IsFalse(registry.RegistryValueExists(registryHive, registryView, registrySubKey, qwordRegistryValueName));
-            Assert.IsFalse(registry.RegistryValueExists(registryHive, registryView, registrySubKey, multiStringRegistryValueName));
-
-            /*
-             * Next, we will test restoring values that did exist when the snapshot was taken, but have since been deleted
-             */
-            // Create a bunch of registry values
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, stringRegistryValueName, stringRegistryValueExpectedValue, stringRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, expandStringRegistryValueName, expandStringRegistryValueExpectedValue, expandStringRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, binaryRegistryValueName, binaryRegistryValueExpectedValue, binaryRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, dwordRegistryValueName, dwordRegistryValueExpectedValue, dwordRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, qwordRegistryValueName, qwordRegistryValueExpectedValue, qwordRegistryValueExpectedKind);
-            registry.SetRegistryValue(registryHive, registryView, registrySubKey, multiStringRegistryValueName, multiStringRegistryValueExpectedValue, multiStringRegistryValueExpectedKind);
-
-            // Simulate taking snapshots of the registry values from another process
-            using (CreateShimsContext())
-            {
-                var snapshotter = CreateSnapshotter();
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, stringRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, expandStringRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, binaryRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, dwordRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, qwordRegistryValueName);
-                snapshotter.SnapshotRegistryValue(registryHive, registryView, registrySubKey, multiStringRegistryValueName);
-            }
-
-            // Delete the registry values
-            registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, stringRegistryValueName);
-            registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, expandStringRegistryValueName);
-            registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, binaryRegistryValueName);
-            registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, dwordRegistryValueName);
-            registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, qwordRegistryValueName);
-            registry.DeleteRegistryValue(registryHive, registryView, registrySubKey, multiStringRegistryValueName);
-
-            // Restore the snapshots
-            restoreSnapshotter.RestoreAbandonedSnapshots();
-
-            // Verify that the registry values have been restored
-            var (stringRegistryValueActualValue, stringRegistryValueActualKind) = registry.GetRegistryValue(registryHive, registryView, registrySubKey, stringRegistryValueName);
-            Assert.AreEqual(stringRegistryValueExpectedValue, stringRegistryValueActualValue);
-            Assert.AreEqual(stringRegistryValueExpectedKind, stringRegistryValueActualKind);
-            var (expandStringRegistryValueActualValue, expandStringRegistryValueActualKind) = registry.GetRegistryValue(registryHive, registryView, registrySubKey, expandStringRegistryValueName);
-            Assert.AreEqual(expandStringRegistryValueExpectedValue, expandStringRegistryValueActualValue);
-            Assert.AreEqual(expandStringRegistryValueExpectedKind, expandStringRegistryValueActualKind);
-            var (binaryRegistryValueActualValue, binaryRegistryValueActualKind) = registry.GetRegistryValue(registryHive, registryView, registrySubKey, binaryRegistryValueName);
-            CollectionAssert.AreEqual(binaryRegistryValueExpectedValue, (byte[])binaryRegistryValueActualValue);
-            Assert.AreEqual(binaryRegistryValueExpectedKind, binaryRegistryValueActualKind);
-            var (dwordRegistryValueActualValue, dwordRegistryValueActualKind) = registry.GetRegistryValue(registryHive, registryView, registrySubKey, dwordRegistryValueName);
-            Assert.AreEqual(dwordRegistryValueExpectedValue, dwordRegistryValueActualValue);
-            Assert.AreEqual(dwordRegistryValueExpectedKind, dwordRegistryValueActualKind);
-            var (qwordRegistryValueActualValue, qwordRegistryValueActualKind) = registry.GetRegistryValue(registryHive, registryView, registrySubKey, qwordRegistryValueName);
-            Assert.AreEqual(qwordRegistryValueExpectedValue, qwordRegistryValueActualValue);
-            Assert.AreEqual(qwordRegistryValueExpectedKind, qwordRegistryValueActualKind);
-            var (multiStringRegistryValueActualValue, multiStringRegistryValueActualKind) = registry.GetRegistryValue(registryHive, registryView, registrySubKey, multiStringRegistryValueName);
-            CollectionAssert.AreEqual(multiStringRegistryValueExpectedValue, (object[])multiStringRegistryValueActualValue);
-            Assert.AreEqual(multiStringRegistryValueExpectedKind, multiStringRegistryValueActualKind);
-        }
-
         #endregion
 
         private RegistrySnapshotter CreateSnapshotter()
